change HSL

+++ WpfApp1/Entity/ConfigData.cs
@@ -50,6 +50,6 @@ namespace WpfApp1.Entity
 
         public string ImageUri { get; set; }
 
-        public int BarLengh { get; set; }
+        public ushort BarLengh { get; set; }
     }
 }
